Hopefully fix compiler-dependent ESb test failure
Apparent "Heisenbug" would not manifest if the value of ptr was
examined inside dec(), either before the conditional or after its
value had been changed, for example, by enabling the DBG macro.

It would also not manifest if compiled without inlined code or if
compiled with clang (6.0.1) instead of GCC (Red Hat 8.2.1-6). It also
manifested on Cygwin with GCC 7.4.0.

The assert should always be true, as either the pointer is null or it
is set to null. Seems to "fix" the bug as a side-effect.